fix: new service to remove circular depedency
